refactor : changed setup based on php-template
