+++ .build/composer.json
@@ -1,0 +1,42 @@
+{
+    "description": "tools",
+    "minimum-stability": "dev",
+    "prefer-stable": true,
+    "require": {
+        "ekino/phpstan-banned-code": "^0.3.1",
+        "felixfbecker/advanced-json-rpc": "^3.0.4",
+        "infection/infection": "^0.15.2",
+        "jwage/changelog-generator": "^1.2.0",
+        "phpstan/phpstan": "^0.12.10",
+        "phpstan/phpstan-deprecation-rules": "^0.12.2",
+        "phpstan/phpstan-mockery": "^0.12.3",
+        "phpstan/phpstan-phpunit": "^0.12.6",
+        "phpstan/phpstan-strict-rules": "^0.12.2",
+        "psalm/plugin-phpunit": "^0.8.1",
+        "slam/phpstan-extensions": "^4.0",
+        "thecodingmachine/phpstan-strict-rules": "^0.12.0",
+        "vimeo/psalm": "^3.8.2",
+        "wikimedia/composer-merge-plugin": "^1.4.1"
+    },
+    "scripts": {
+        "cs": "php-cs-fixer fix --config=\"./../.php_cs\"",
+        "cs:check": "php-cs-fixer fix --config=\"./../.php_cs\" --dry-run",
+        "phpstan": "phpstan analyse -c ./../phpstan.neon",
+        "psalm": "psalm --threads=$(nproc)",
+        "psalm:fix": "psalm --alter --issues=all --threads=$(nproc)",
+        "infection": "infection --configuration=\"./../infection.json\" -j$(nproc)"
+    },
+    "extra": {
+        "prefetcher": {
+            "require": {
+                "phpunit/phpunit": "^8.5",
+                "friendsofphp/php-cs-fixer": "^2.16.0"
+            }
+        },
+        "merge-plugin": {
+            "include": [
+                "../composer.json"
+            ]
+        }
+    }
+}
