+++ .gitattributes
@@ -1,18 +1,17 @@
 * text=auto
 *.php text eol=lf
 
-build              export-ignore
+.dependabot        export-ignore
 .github            export-ignore
+build              export-ignore
 
-.php_cs            export-ignore
+.editorconfig      export-ignore
 .gitattributes     export-ignore
 .gitignore         export-ignore
+.php_cs            export-ignore
 .travis.yml        export-ignore
-.editorconfig      export-ignore
+CODE_OF_CONDUCT.md export-ignore
+CONTRIBUTING.md    export-ignore
 phpstan.neon       export-ignore
-
-phpunit.xml.dist               export-ignore
-
-CONTRIBUTING.md                export-ignore
-README.md                      export-ignore
-CODE_OF_CONDUCT.md             export-ignore
+phpunit.xml.dist   export-ignore
+README.md          export-ignore
