+++ .github/CONTRIBUTING.md
@@ -1,0 +1,24 @@
+# Contributing
+Narrowspark and Narrowspark package are open source, community-driven projects.
+
+If you'd like to contribute, please read the following documents:
+
+* [Reviewing issues/pull requests][0]
+* [Reporting a Bug][1]
+* [Submitting a Patch][2]
+* [Narrowspark Core Team][3]
+* [Security Issues][4]
+* [Running Narrowspark Tests][5]
+* [Our Backwards Compatibility Promise][6]
+* [Coding Standards][7]
+* [Conventions][8]
+
+[0]: https://narrowspark.com/docs/current/contributing/community/reviews.html
+[1]: https://narrowspark.com/docs/current/contributing/code/bugs.html
+[2]: https://narrowspark.com/docs/current/contributing/code/patches.html
+[3]: https://narrowspark.com/docs/current/contributing/code/core_team.html
+[4]: https://narrowspark.com/docs/current/contributing/code/security.html
+[5]: https://narrowspark.com/docs/current/contributing/code/tests.html
+[6]: https://narrowspark.com/docs/current/contributing/code/bc.html
+[7]: https://narrowspark.com/docs/current/contributing/code/standards.html
+[8]: https://narrowspark.com/docs/current/contributing/code/conventions.html
