+++ .github/ISSUE_TEMPLATE/Bug_report.md
@@ -1,0 +1,23 @@
+---
+name: 🐛 Bug Report
+about: If something isn't working as expected.
+ 
+---
+ 
+<!--
+    The Code of Conduct (../CODE_OF_CONDUCT.md) applies to all the activity on this repository.
+-->
+ 
+**Narrowspark version(s) affected**: x.y.z
+
+**Description**
+<!-- A clear and concise description of the problem. -->
+ 
+**How to reproduce**
+<!--- Code and/or config needed to reproduce the problem. -->
+ 
+**Possible Solution**
+<!-- Optional: only if you have suggestions on a fix/reason for the bug -->
+ 
+**Additional context**
+<!-- Optional: any other context about the problem: log messages, screenshots, etc. -->
