+++ .github/ISSUE_TEMPLATE/Documentation_issue.md
@@ -1,0 +1,9 @@
+---
+name: 📚 Documentation Issue
+about: Anything related to Narrowspark package documentation
+ 
+---
+ 
+<!--
+    The Code of Conduct (../CODE_OF_CONDUCT.md) applies to all the activity on this repository.
+-->
