+++ .github/ISSUE_TEMPLATE/Feature_request.md
@@ -1,0 +1,21 @@
+---
+name: 🚀 Feature Request
+about: I have a suggestion (and might want to implement myself 🙂)!
+
+---
+
+<!--
+    The Code of Conduct (../CODE_OF_CONDUCT.md) applies to all the activity on this repository.
+-->
+
+**Is your feature request related to a problem? Please describe.**
+<!-- A clear and concise description of what the problem is. Ex. I have an issue when [...] -->
+
+**Describe the solution you'd like**
+<!-- A clear and concise description of what you want to happen. Add any considered drawbacks. -->
+
+**Describe alternatives you've considered**
+<!-- A clear and concise description of any alternative solutions or features you've considered. -->
+
+**Teachability, Documentation, Adoption, Migration Strategy**
+<!-- If you can, explain how users will be able to use this and possibly write out a version the docs. -->
