+++ .github/ISSUE_TEMPLATE/Security_issue.md
@@ -1,0 +1,15 @@
+---
+name: ㊙️ Security Issue
+about: Report security-related errors
+ 
+---
+ 
+<!--
+    The Code of Conduct (../CODE_OF_CONDUCT.md) applies to all the activity on this repository.
+-->
+
+⚠ PLEASE DON'T DISCLOSE SECURITY-RELATED ISSUES PUBLICLY, SEE BELOW.
+ 
+If you have found a security issue in Narrowspark, please send the details to
+security [at] narrowspark.com and don't disclose it publicly until we can provide a
+fix for it.
