+++ .github/ISSUE_TEMPLATE/Support_question.md
@@ -1,0 +1,16 @@
+---
+name: 🤗 Support Question
+about: If you have a question 💬, please check out our [Discord](https://discord.gg/UBG6r9b) or [StackOverflow](https://stackoverflow.com/)!
+
+---
+ 
+<!--
+    The Code of Conduct (../CODE_OF_CONDUCT.md) applies to all the activity on this repository.
+-->
+
+Issues on GitHub are intended to be related to problems with Narrowspark packages itself, so we recommend not using this medium to ask them here 😁. 
+
+For this kind of questions about using Narrowspark or third-party packages, please use
+any of the support alternatives.
+ 
+Thanks!
