+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,0 +1,22 @@
+| Q             | A
+| ------------- | ---
+| Branch?       | master   <!-- for features -->
+| Bug fix?      | yes/no
+| New feature?  | yes/no   <!-- don't forget to update CHANGELOG.md file(s) -->
+| Deprecations? | yes/no   <!-- don't forget to update UPGRADE-*.md file(s) -->
+| Tickets       | Fix #... <!-- prefix each issue number with "Fix #", if any -->
+| License       | MIT
+| Doc PR        | yes/no
+
+<!--
+Replace this notice by a short README for your feature/bugfix. This will help people
+understand your PR and can be used as a start for the documentation.
+
+Additionally:
+ - Always add tests and ensure they pass.
+ - Never break backward compatibility
+ - Bug fixes must be submitted against the lowest branch where they apply
+   (lowest branches are regularly merged to upper ones so they get the fixes too).
+ - Features and deprecations must be submitted against the branch "master".
+ - Legacy code removals go to the master branch.
+-->
