+++ .github/workflows/greetings.yml
@@ -1,0 +1,15 @@
+# https://help.github.com/en/categories/automating-your-workflow-with-github-actions
+
+name: "Greetings"
+
+on: ["pull_request", "issues"]
+
+jobs:
+  greeting:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/first-interaction@v1
+        with:
+          repo-token: ${{ secrets.GITHUB_TOKEN }}
+          issue-message: 'Awesome! Thank you for taking the time to create your first issue! Please review the [guidelines](https://narrowspark.com/docs/current/contributing)'
+          pr-message: 'Great! Thank you for taking the time to create your first pull request! Please review the [guidelines](https://narrowspark.com/docs/current/contributing)'
