+++ .github/workflows/markdown-lint.yml
@@ -1,0 +1,31 @@
+# https://help.github.com/en/categories/automating-your-workflow-with-github-actions
+
+name: "Markdown Lint"
+
+on:
+  push:
+    paths:
+      - "**.md"
+  pull_request:
+    paths:
+      - "**.md"
+
+jobs:
+  text-lint:
+    name: "Markdown lint"
+
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v2
+        with:
+          fetch-depth: 2
+
+      - name: "Setup node"
+        uses: actions/setup-node@v1
+        with:
+          node-version: "10.16.0"
+
+      - run: npm install
+
+      - run: npm run lint
