+++ .nvmrc
@@ -1,0 +1,1 @@
+10.16.0
