+++ .php_cs
@@ -1,19 +1,37 @@
 <?php
+
 declare(strict_types=1);
+
+use Ergebnis\License;
 use Narrowspark\CS\Config\Config;
 
-$config = new Config();
+$license = License\Type\MIT::markdown(
+    __DIR__ . '/LICENSE.md',
+    License\Range::since(
+        License\Year::fromString('2015'),
+        new \DateTimeZone('UTC')
+    ),
+    License\Holder::fromString('Daniel Bannert'),
+    License\Url::fromString('https://github.com/narrowspark/php-cs-fixer-config')
+);
+
+$license->save();
+
+$config = new Config($license->header(), []);
+
 $config->getFinder()
     ->files()
     ->in(__DIR__)
-    ->exclude('build')
-    ->exclude('vendor')
+    ->exclude([
+        'build',
+        'vendor',
+        '.dependabot',
+        '.github',
+    ])
     ->name('*.php')
     ->ignoreDotFiles(true)
     ->ignoreVCS(true);
 
-$cacheDir = getenv('TRAVIS') ? getenv('HOME') . '/.php-cs-fixer' : __DIR__;
-
-$config->setCacheFile($cacheDir . '/.php_cs.cache');
+$config->setCacheFile(__DIR__ . '/.build/php-cs-fixer/.php_cs.cache');
 
 return $config;
