+++ .textlintrc
@@ -1,0 +1,34 @@
+{
+  "filters": {},
+  "rules": {
+    "en-capitalization": true,
+    "footnote-order": true,
+    "no-todo": true,
+    "no-dead-link": {
+        "ignore": [
+            "bc_data_*",
+        ]
+    },
+    "no-empty-section": true,
+    "terminology": true,
+    "apostrophe": true,
+    "diacritics": true,
+    "@textlint-rule/no-invalid-control-character": true,
+    "@textlint-rule/no-unmatched-pair": true,
+    "abbr-within-parentheses": true,
+    "alex": {
+        "allow": [
+            "period",
+            "european",
+            "failure",
+            "fore",
+            "attack",
+            "execution",
+            "executed",
+            "remain",
+            "execute"
+        ]
+    },
+    "@textlint-rule/preset-google": true
+  }
+}
