+++ README.md
@@ -1,9 +1,8 @@
 <h1 align="center">Narrowspark php-cs-fixer Config</h1>
 <p align="center">
     <a href="https://github.com/narrowspark/php-cs-fixer-config/releases"><img src="https://img.shields.io/packagist/v/narrowspark/php-cs-fixer-config.svg?style=flat-square"></a>
-    <a href="https://php.net/"><img src="https://img.shields.io/badge/php-%5E7.1.0-8892BF.svg?style=flat-square"></a>
+    <a href="https://php.net/"><img src="https://img.shields.io/badge/php-%5E7.3.0-8892BF.svg?style=flat-square"></a>
     <a href="https://codecov.io/gh/narrowspark/php-cs-fixer-config"><img src="https://img.shields.io/codecov/c/github/narrowspark/php-cs-fixer-config/master.svg?style=flat-square"></a>
-    <a href="https://gitter.im/narrowspark/framework"><img src="https://img.shields.io/gitter/room/nwjs/nw.js.svg?style=flat-square"></a>
     <a href="https://opensource.org/licenses/MIT"><img src="https://img.shields.io/badge/license-MIT-brightgreen.svg?style=flat-square"></a>
 </p>
 This repository provides a configuration for https://github.com/FriendsOfPHP/PHP-CS-Fixer, which
@@ -32,7 +31,7 @@ $config = new Config();
 $config->getFinder()
     ->files()
     ->in(__DIR__)
-    ->exclude('build')
+    ->exclude('.build')
     ->exclude('vendor')
     ->name('*.php')
     ->ignoreDotFiles(true)
@@ -77,7 +76,7 @@ $config = new Config(null /* if you dont need a header */, [
 $config->getFinder()
     ->files()
     ->in(__DIR__)
-    ->exclude('build')
+    ->exclude('.build')
     ->exclude('vendor')
     ->name('*.php')
     ->ignoreDotFiles(true)
