+++ composer.json
@@ -20,19 +20,16 @@
         }
     ],
     "require": {
-        "php": "^7.2",
+        "php": "^7.3",
+        "ergebnis/license": "~0.1.0",
         "friendsofphp/php-cs-fixer": "~2.16.1",
         "kubawerlos/php-cs-fixer-custom-fixers": "~1.17.0",
         "pedrotroller/php-cs-custom-fixer": "~2.19.1"
     },
     "require-dev": {
-        "narrowspark/testing-helper": "^8.0.2",
-        "phpstan/phpstan": "^0.12.10",
-        "phpstan/phpstan-deprecation-rules": "^0.12.2",
-        "phpstan/phpstan-phpunit": "^0.12.6",
-        "phpstan/phpstan-strict-rules": "^0.12.2",
+        "ext-json": "*",
         "phpunit/phpunit": "^8.5.2",
-        "thecodingmachine/phpstan-strict-rules": "^0.12.0"
+        "narrowspark/testing-helper": "^8.0.2"
     },
     "config": {
         "sort-packages": true
@@ -50,13 +47,29 @@
     "minimum-stability": "dev",
     "prefer-stable": true,
     "scripts": {
-        "coverage": "phpunit --coverage-html=\"build/logs\"",
-        "cs": "php-cs-fixer fix -vvv",
-        "phpstan": "phpstan analyse -c phpstan.neon src --memory-limit=-1",
-        "test": "phpunit"
+        "coverage": [
+            "phpunit --dump-xdebug-filter=./.build/phpunit/.xdebug-filter.php",
+            "phpunit --prepend=./.build/phpunit/.xdebug-filter.php --coverage-html=./build/phpunit/coverage"
+        ],
+        "cs": "composer --working-dir=./.build cs -- -v",
+        "cs:check": "composer --working-dir=./.build cs:check -- -v",
+        "phpstan": "composer --working-dir=./.build phpstan -- --memory-limit=-1",
+        "psalm": "composer --working-dir=./.build psalm",
+        "psalm:fix": "composer --working-dir=./.build psalm:fix",
+        "infection": "composer --working-dir=./.build infection -- --min-covered-msi=100 --min-msi=100",
+        "test": "phpunit",
+        "post-install-cmd": "composer --working-dir=./.build install",
+        "post-update-cmd": "composer --working-dir=./.build update --lock"
     },
     "support": {
         "issues": "https://github.com/narrowspark/php-cs-fixer-config/issues",
         "source": "https://github.com/narrowspark/php-cs-fixer-config"
+    },
+    "extra": {
+        "prefetcher": {
+            "require": {
+                "phpunit/phpunit": "^8.5"
+            }
+        }
     }
 }
