+++ docker-compose.yml
@@ -1,0 +1,22 @@
+version: '3'
+services:
+    php73_64:
+      build:
+        context: ./.docker/php/
+        args:
+          NAMESPACE: "php"
+          PHP_VERSION: "7.3"
+          XDEBUG_VERSION: "2.9.0"
+          INSTALL_XDEBUG: "true"
+      expose:
+        - 9000
+      volumes:
+        - data:/var/www/package
+      tty: true
+
+volumes:
+  data:
+    driver: lebokus/bindfs:latest
+    driver_opts:
+      sourcePath: "${PWD}"
+      map: "${UID:-1000}/0:@${UID:-1000}/@0"
