+++ infection.json
@@ -1,0 +1,23 @@
+{
+    "timeout": 10,
+    "source": {
+        "directories": [
+            "./../src"
+        ]
+    },
+    "logs": {
+        "text": "infection/infection-log.txt"
+    },
+    "phpUnit": {
+        "configDir": "./",
+        "customPath": "./vendor/phpunit/phpunit/phpunit"
+    },
+    "mutators": {
+        "@default": true,
+        "FalseValue": false,
+        "TrueValue": false,
+        "IncrementInteger": false,
+        "DecrementInteger": false,
+        "ArrayItemRemoval": false
+    }
+}
